requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.6.0
async-timeout==4.0.3
attrs==25.1.0
autobahn==23.1.2
autoflake==2.3.1
Automat==24.8.1
autopep8==2.3.1
Babel==2.10.1
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.5.18.1
cffi==1.17.1
cfgv==3.4.0
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.3.2
click==8.1.3
cloudinary==1.41.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
dataclasses-json==0.6.7
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dj-rest-auth==5.0.2
Django==4.2.19
django-allauth==0.58.2
django-cors-headers==3.13.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-money==2.1.1
django-quill-editor==0.1.40
django-shell-plus==1.1.7
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.2.2
factory_boy==3.3.1
faiss-cpu==1.8.0.post1
Faker==30.1.0
filelock==3.16.1
flake8==7.1.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
identify==2.6.1
idna==3.10
importlib-metadata==4.11.4
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
Markdown==3.3.7
marshmallow==3.22.0
mccabe==0.7.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==0.4.3
nodeenv==1.9.1
numpy==1.24.4
oauthlib==3.2.2
openai==1.68.2
orjson==3.10.15
packaging==24.1
pathspec==0.9.0
pillow==10.4.0
platformdirs==2.5.2
pluggy==1.5.0
pre-commit==3.5.0
propcache==0.2.0
psycopg2==2.9.3
py-moneyed==1.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sentry-sdk==2.14.0
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.4.2
stripe==10.1.0
tenacity==8.5.0
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.21.1
yarl==1.15.2
zipp==3.8.0
zope.interface==7.2
